factor out arrays, flags
