+++ tests/flags.witx
@@ -1,0 +1,16 @@
+(use "errno.witx")
+
+(typename $car_config
+  (flags u8
+    $automatic
+    $awd
+    $suv))
+
+(module $flags
+  (@interface func (export "configure_car")
+    (param $old_config $car_config)
+    (param $old_config_by_ptr (@witx const_pointer $car_config))
+    (result $error $errno)
+    (result $new_config $car_config)
+  )
+)
